requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
black==20.8b1
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
distlib==0.3.1
dj-database-url==0.5.0
Django==2.2.10
django-crontab==0.7.1
django-jet==1.0.8
django-phonenumber-field==5.0.0
django-twilio==0.13.0.post1
djangorestframework==3.11.1
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
install==1.3.3
isort==5.5.2
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.0.3
numpy==1.19.2
pandas==1.1.2
pathspec==0.8.0
phonenumbers==8.12.9
Pillow==7.2.0
prettytable==0.7.2
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
pyotp==2.4.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
regex==2020.7.14
requests==2.24.0
schedule==0.6.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
twilio==6.32.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
Unipath==1.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
whitenoise==5.2.0
wrapt==1.12.1
xlwt==1.3.0
zipp==3.1.0
